+++ test/unit/object/test_file.py
@@ -580,6 +580,7 @@ def test_get_shared_link_download_url(
         shared_link_unshared_at,
         shared_link_password,
         shared_link_can_preview,
+        shared_link_vanity_name,
         test_url,
         etag,
         if_match_header,
@@ -606,12 +607,16 @@ def test_get_shared_link_download_url(
         permissions['can_preview'] = shared_link_can_preview
     if shared_link_password is not None:
         expected_data['shared_link']['password'] = shared_link_password
+    if shared_link_vanity_name is not None:
+        expected_data['shared_link']['vanity_name'] = shared_link_vanity_name
+
     url = test_file.get_shared_link_download_url(
         etag=etag,
         access=shared_link_access,
         unshared_at=shared_link_unshared_at,
         password=shared_link_password,
         allow_preview=shared_link_can_preview,
+        vanity_name=shared_link_vanity_name,
     )
     mock_box_session.put.assert_called_once_with(
         expected_url,